src/u2net_pytorch/requirements_train.txt
# -f https://data.pyg.org/whl/torch-1.10.0+cu102.html
# -f https://data.pyg.org/whl/torch-1.10.0+cu111.html
python-dotenv
mlflow
opencv-python
pytorch-lightning
